move energy & fault updates to main loop
There are 2 loops which gather data from RAPI;
handleRapiRead and update_rapi_values.  Not super clearly
named, but the first is run only once at startup, while the
other is run every 5s from the main loop.

Energy use & fault information may change over time, so
they should be updated in the main loop.  Without this
they are red once at startup and never changed.

The downside here is that every 5s one of the (now 6)
commands is run, so it takes 30s to get every one of the
6 RAPI commands updated.

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>